refactor: rename classes and modules
