+++ src/docformatter/__init__.py
@@ -28,10 +28,10 @@ __all__ = ["__version__"]
 # docformatter Local Imports
 from .__pkginfo__ import __version__
 from .strings import *
-from .syntaxor import *
+from .syntax import *
 from .util import *
 
 # Have isort skip these they require the functions above.
-from .configurator import Configurator  # isort: skip
-from .encodor import Encodor  # isort: skip
-from .formattor import Formator, FormatResult  # isort: skip
+from .configuration import Configurater  # isort: skip
+from .encode import Encoder  # isort: skip
+from .format import Formatter, FormatResult  # isort: skip
