+++ src/docformatter/syntax.py
