+++ tests/test_configuration_functions.py
@@ -25,7 +25,7 @@
 # ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-"""Module for testing docformatter's Configurator class."""
+"""Module for testing docformatter's Configurater class."""
 
 # Standard Library Imports
 import io
@@ -35,21 +35,21 @@ import sys
 import pytest
 
 # docformatter Package Imports
-from docformatter import Configurator
+from docformatter import Configurater
 
 
-class TestConfigurator:
+class TestConfigurater:
     """Class for testing configuration functions."""
 
     @pytest.mark.unit
     def test_initialize_configurator_with_default(self):
-        """Return a Configurator() instance using default pyproject.toml."""
+        """Return a Configurater() instance using default pyproject.toml."""
         argb = [
             "/path/to/docformatter",
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.args_lst == argb
@@ -57,7 +57,7 @@ class TestConfigurator:
 
     @pytest.mark.unit
     def test_initialize_configurator_with_pyproject_toml(self):
-        """Return a Configurator() instance loaded from a pyproject.toml."""
+        """Return a Configurater() instance loaded from a pyproject.toml."""
         argb = [
             "/path/to/docformatter",
             "-c",
@@ -66,7 +66,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.args.check
@@ -94,7 +94,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.config_file == "./tests/_data/setup.cfg"
@@ -115,7 +115,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.config_file == "./tests/_data/tox.ini"
@@ -136,7 +136,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.config_file == "./tests/conf.py"
@@ -157,7 +157,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.config_file == "./tests/_data/tox.ini"
@@ -193,7 +193,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.args.line_range == [1, 3]
@@ -210,7 +210,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         with pytest.raises(SystemExit):
             uut.do_parse_arguments()
 
@@ -230,7 +230,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         with pytest.raises(SystemExit):
             uut.do_parse_arguments()
 
@@ -253,7 +253,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         uut.do_parse_arguments()
 
         assert uut.args.length_range == [25, 55]
@@ -270,7 +270,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         with pytest.raises(SystemExit):
             uut.do_parse_arguments()
 
@@ -290,7 +290,7 @@ class TestConfigurator:
             "",
         ]
 
-        uut = Configurator(argb)
+        uut = Configurater(argb)
         with pytest.raises(SystemExit):
             uut.do_parse_arguments()
